Move User in gormungandr to avoid circular dependencies
